+++ src/display/texture.h
@@ -13,7 +13,7 @@ class Texture
 public:
   Texture(void) = default;
   Texture(std::string const& file, class Window& window);
-  Texture(std::string const& text, sdl::Font const& font, int width, class Window& window);
+  Texture(std::string const& text, sdl::Font const& font, float width, class Window& window);
   Texture(std::string const& text, sdl::Font const& font, class Window& window);
   Texture(Texture const& texture, BoundingBox const& clip);
   bool operator()(BoundingBox const& source, BoundingBox const& destination, float parallax, bool tile, double angle, Modulation const& modulation) const;
